Fix POM editor bug with XML namespaces
It seems that the XML parser position reporting has troubles when the file
contains namespaces. Although it is not usual for a POM file not to use
the default namespace only, it might be troublesome.

Unfortunately, just a heuristic work-around could be applied: the editor
tries to find the boundaries of the text node based on the given offset,
even for the start which is accurate otherwise.